Fix kernel deadlock on open/close
In 648eabb5bda05f7bc3d15eb5cfce92f60b6fd718 a fix was added to remove
a race condition between the creation and desruction of a proc entry
for a thread group. The issue was caused by a user code opening then
closing the xpmem device file before using xpmem. Unfortunately the
fix causes deadlock in 4.x kernels (not observed in 3.x). This commit
removes the open spinlock and moves the destruction of the proc entry
into xpmem_flush. This will guarantee that the proc entry is always
destroyed before it is created again.

An additional change introduced in this commit has to do with when the
thread group is removed from the hash table. Before this commit the
hash table read lock is obtained, the returned tg entry referenced,
and the read lock dropped. The write lock is later obtained and the
entry is removed. This allows for the possibility of a race
condition. To avoid this a non-locking version of
xpmem_tg_ref_by_tgid_all_nolock has been implemented to allow a search
then remove functionality.

Fixes hjelmn/xpmem#4

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>