Fix bug in munmap workaround
The original Cray code used the unexported do_munmap function to
specify the memory descriptor where the address range should be
unmapped. To avoid the use of a unexported symbol the code was updated
to use the vm_munmap function instead. In order for this to work it
was necessary to set the current task's mm value before calling
vm_munmap. This appears to cause hangs, crashes, etc in 4.x kernels
that were not observed with 3.x. To fix the issue remove the set of
current->mm and just avoid calling vm_munmap when the current task has
no memory descriptor. The memory mapping should go away automatically
when the last mmput is called on the memory descriptor. Tested with
4.4.1 and confirmed there are no apparent leaks.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>