libxpmem: cleanup libxpmem and update license
This commit moves the libxpmem internal portions of xpmem.h (ioctl and
command structure definitions) into a seperate header
(xpmem_internal.h). The doxygen documentation of libxpmem functions
has been moved the xpmem.h and the use of the reserved identifier
_XPMEM_H has been removed.

This commit also updates the license of just libxpmem to LGPL. This
change was approved by the upstream author (Cray Inc). The license on
the headers xpmem.h and xpmem_internal.h remain GPL at this time. This
should be sufficient to allow distribution of non-GPL software linked
against libxpmem.

Fixes #6

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>