xpmem_detach: fix deadlock
I made modifications to the detach code to use vm_munmap instead of
the non-exported do_munmap call. This required dropping the mmap
semaphore before calling vm_munmap. Unfortunately this causes hold and
wait with xpmem_clear_PTEs_of_att. If the detach thread goes to sleep
between dropping the semaphore re-acquiring it and a thread in
xpmem_clear_PTEs_of_att acquires the semaphore then we have one thread
with the mmap semaphore trying to get the attachment lock and another
thread with the attachement lock trying to get the mmap
semaphore. Fixed the issue by clearing the attachment state and
dropping the attachment lock before dropping the mmap semaphore.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>