install module in <prefix>/lib/modules/<linux_release>
This commit changes the directory where the xpmem module is. If no
kernel source directory is specified (build for current running
release) then linux_release is set from uname -r. If a kernel source
directory is specified the release is taken from the directory name if
the source specified is /lib/modules/release/build or
/lib/modules/release/source or from make kernel_release otherwise.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>